Use exec to replace the shell
I'm not a huge fan of replacing the shell by modifying `.bashrc`. I don't think this is really having a huge impact- the biggest one is that `exit` (or ctrl+d, etc.) will behave unexpectedly if that is done. Using `exec` would help, although I'm not sure that's a great solution either.