New Tip for argo cd
When you copy paste in the shell script. It sometimes shortens the script and displays "...." at the beginning in the terminal. When you hit return it fills in the whole script and runs correctly. The Tip is there to reassure the student that it is not a copy paste error.