🌈 update estimation.md 🌈
seen something in the estimation slides that doesn’t seem right to me, which is in the “How do we do it slide” it says: “First, the product backlog refinement and the sprint planning must be done” - which to me, the use of the word “must” means that it is a delivery practice. However, as this is a foundation practice, I think that the wording of this should be changed to accommodate doing it outside the delivery phase. So I updated the wording on the slide